rpki-client: use ASN1_BIT_STRING_get_length()
Instead of fiddling with the undocumented ASN1_STRING_FLAG_BITS_LEFT
in the flags of an ASN.1 bit string, convert to the API outlined by
beck in [1].

For IP addresses like in RFC 3779, RFC 9582, etc, this is entirely
straightforward and results in simpler and more readable code.

For manifests, the hash in a fileHash is a SHA-256 encoded in a
32-octet BIT STRING for whatever reason, so we should really be
checking that the unused octet is 0. We punted on this because
the flag wrangling got a bit too much. With the API it is simple.

[1]: https://github.com/openssl/openssl/issues/29184